+++ ts/src/core/query/custom_query_sqlite.test.ts
@@ -6,16 +6,12 @@ import {
 } from "../../testutils/fake_data/index";
 import { commonTests } from "./shared_test";
 
-beforeEach(async () => {
-  QueryRecorder.clear();
-});
-
 commonTests({
   newQuery(viewer: Viewer, user: FakeUser) {
     return UserToContactsFkeyQuery.query(viewer, user);
   },
   tableName: "fake_contacts",
-  where: "user_id = $1",
+  where: "user_id = ?",
   sortCol: "created_at",
-  livePostgresDB: true, // doing this on a db as opposed to in memory
+  sqlite: true,
 });
